+++ src/include/sll/debug.h
@@ -5,6 +5,10 @@
 
 
 
+__SLL_FUNC void sll_debug_print_assembly(const sll_assembly_data_t* a_dt);
+
+
+
 __SLL_FUNC void sll_debug_print_object(const sll_object_t* o,const sll_compilation_data_t* c_dt);