+++ src/include/sll/integer_list.h
@@ -9,15 +9,19 @@
 
 
 
-__SLL_FUNC void sll_free_integer_list(sll_integer_list_t* o);
+__SLL_FUNC void sll_free_integer_list(sll_integer_list_t* il);
 
 
 
-__SLL_FUNC sll_bool_t sll_integer_list_add(sll_integer_list_t* o,sll_integer_t v);
+__SLL_FUNC sll_bool_t sll_integer_list_add(sll_integer_list_t* il,sll_integer_t k,void* v);
 
 
 
-__SLL_FUNC sll_bool_t sll_integer_list_contains(sll_integer_list_t* o,sll_integer_t v);
+__SLL_FUNC sll_bool_t sll_integer_list_contains(sll_integer_list_t* il,sll_integer_t k);
+
+
+
+__SLL_FUNC sll_bool_t sll_integer_list_get(sll_integer_list_t* il,sll_integer_t k,void** v);